Update coroutine commands to be inline with current code styling/patterns
